Better extract platform independent code
